// Script Modificado/illenium-appearance/web/src/components/Appearance/interfaces.ts
export interface PedComponent {
  component_id: number;
  drawable: number;
  texture: number;
}
// ...
export interface ComponentSettings {
  component_id: number;
  drawable: {
    min: number;
    max: number;
  };
  texture: {
    min: number;
    max: number;
  };
  blacklist: BlacklistSettings;
}

interface BlacklistSettings {
  drawables: number[];
  textures: number[];
}
// ...
export interface ComponentConfig {
  masks: boolean;
  upperBody: boolean;
  lowerBody: boolean;
  bags: boolean;
  shoes: boolean;
  scarfAndChains: boolean;
  shirts: boolean;
  bodyArmor: boolean;
  decals: boolean;
  jackets: boolean;
}